extract DocumentEdits utils
