+++ Source/DafnyLanguageServer.Test/CodeActions/CodeActionsTest.cs
@@ -310,8 +310,8 @@ function Foo(i: int): int
               codeAction = await RequestResolveCodeAction(codeAction);
               var textDocumentEdit = codeAction.Edit?.DocumentChanges?.Single().TextDocumentEdit;
               Assert.NotNull(textDocumentEdit);
-              var modifiedOutput = string.Join("\n", ApplyEdits(textDocumentEdit, output)).Replace("\r\n", "\n");
-              var expectedOutput = string.Join("\n", ApplySingleEdit(ToLines(output), expectedRange, expectedNewText)).Replace("\r\n", "\n");
+              var modifiedOutput = string.Join("\n", DocumentEdits.ApplyEdits(textDocumentEdit, output)).Replace("\r\n", "\n");
+              var expectedOutput = string.Join("\n", DocumentEdits.ApplySingleEdit(DocumentEdits.ToLines(output), expectedRange, expectedNewText)).Replace("\r\n", "\n");
               Assert.Equal(expectedOutput, modifiedOutput);
             }
           }
@@ -323,34 +323,6 @@ function Foo(i: int): int
     }
 
     public CodeActionTest(ITestOutputHelper output) : base(output) {
-    }
-
-    private static List<string> ApplyEdits(TextDocumentEdit textDocumentEdit, string output) {
-      var inversedEdits = textDocumentEdit.Edits.ToList()
-        .OrderByDescending(x => x.Range.Start.Line)
-        .ThenByDescending(x => x.Range.Start.Character);
-      var modifiedOutput = ToLines(output);
-      foreach (var textEdit in inversedEdits) {
-        modifiedOutput = ApplySingleEdit(modifiedOutput, textEdit.Range, textEdit.NewText);
-      }
-
-      return modifiedOutput;
-    }
-
-    private static List<string> ToLines(string output) {
-      return output.ReplaceLineEndings("\n").Split("\n").ToList();
-    }
-
-    private static List<string> ApplySingleEdit(List<string> modifiedOutput, Range range, string newText) {
-      var lineStart = modifiedOutput[range.Start.Line];
-      var lineEnd = modifiedOutput[range.End.Line];
-      modifiedOutput[range.Start.Line] =
-        lineStart.Substring(0, range.Start.Character) + newText +
-        lineEnd.Substring(range.End.Character);
-      modifiedOutput = modifiedOutput.Take(range.Start.Line).Concat(
-        modifiedOutput.Skip(range.End.Line)
-      ).ToList();
-      return modifiedOutput;
     }
   }
 }
